jGnoetry has been refactored to remove almost all depence on the html form from the generation routine. The actual referenceing of the text-buttons has yet to be refactored. Naming conventions could be better. jQuery is probably not necessary -- and could be refactored out as well; it was originally added for a purpose that turned out not to be needed, but by the time that was discovered, it had been used for a lot of initialization.
I tried to separate conconcerns into separate files -- default texts,
gui.js, css file, and the jGnoetry.js core code.

I did major changes to control.html by extracting (almost?) all of the inline-css and all of the inline javascript.
Rendered html still has hard-coded javascript and magic-strings-as-dom elements;
I don't know if that will be changed down the line.
I did some minor tweaking to control.html layout for consistency of code -- and display, I felt.

I would prefer to use divs and css instead of frames, but that will only be tackled
after the last jGnoetry.js direct dependence on the .html files is removed.

The goal for all of this work was to pare down the core code, to make it more obvious what was poetry generation
by making it separate from presentation and options-gathering.

gui.js is called from onDocumentReady, and takes care of.. gui elements.
hiding, showing, attaching javascript to controls.
functions that are directly called from controls that change the display, or initiate an action.
Additionally, this holds the options-gathering code that extracts settings from control.html
and passes them to the core code as required.

defaultTexts.js holds the large default texts
when they were inlined into the html, they were actually easier to deal with.
but now, they are separate, so the html and other code is easier to deal with.

jGnoetry.css holds display settings - selected/not-selected elements are still hard-coded as non-semantic styles.

jGnoetry.js is the core code of poetry-generation.